+++ behave_retry/__init__.py
@@ -7,7 +7,7 @@ from .exceptions import RetryExhaustedError
 from .hooks import after_scenario_hook, retry_report, setup_retry
 from .stats import RetryStats, ScenarioRetry
 
-__version__ = "1.0.0"
+__version__ = "1.0.1"
 
 __all__ = [
     "setup_retry",